refactor(VM-618): extract hooks CLI to cli_commands/claude.py module
Move hooks CLI code from cli.py (3848 lines) to its proper home in
cli_commands/claude.py (378 lines), matching the existing pattern used
by agent.py, status.py, exchanges.py and other CLI command modules.
